+++ src/main/php/xp/reflection/ReflectRunner.class.php
@@ -42,6 +42,11 @@ class ReflectRunner {
     foreach ($args as $arg) {
       if ('-a' === $arg || '--all' === $arg) {
         $flags |= Information::ALL;
+      } else if ('-d' === $arg || '--doc' === $arg) {
+        $flags |= Information::DOC;
+      } else if ('' === $arg || '-' === $arg[0]) {
+        Console::$err->writeLine('Error: unknown argument "'.$arg.'"');
+        return 1;
       } else {
         $name= $arg;
       }
@@ -51,11 +56,11 @@ class ReflectRunner {
       Console::$err->writeLine('Error: no class or package name given');
       return 1;
     } else if (strstr($name, \xp::CLASS_FILE_EXT)) {
-      $information= Information::forClass($cl->loadUri(realpath($name)));
+      $information= Information::forClass($cl->loadUri(realpath($name)), $flags);
     } else if ($cl->providesClass($name)) {
-      $information= Information::forClass($cl->loadClass($name));
+      $information= Information::forClass($cl->loadClass($name), $flags);
     } else if ($cl->providesPackage($name)) {
-      $information= Information::forPackage($name);
+      $information= Information::forPackage($name, $flags);
     } else {
       Console::$err->writeLine('Error: "'.$name.'" neither refers to an existing class or package');
       return 2;
@@ -64,13 +69,13 @@ class ReflectRunner {
     foreach ($information->sources() as $source) {
       Console::writeLine("\e[33m@", $source, "\e[0m");
     }
-
-    $information->display($flags, new WithHighlighting(Console::$out, [
+    $information->display(new WithHighlighting(Console::$out, [
       '/(class|enum|trait|interface|package|directory|function) (.+)/'      => "\e[34m\$1\e[0m \$2",
       '/(extends|implements) ([^ ]+)/'                                      => "\e[34m\$1\e[0m \$2",
       '/\b(var|int|string|float|array|iterable|object|void|static|self)\b/' => "\e[36m\$1\e[0m",
-      '/(public|private|protected|abstract|final|static)/'                  => "\e[1;35m\$1\e[0m",
+      '/(public|private|protected|abstract|final|static) /'                 => "\e[1;35m\$1 \e[0m",
       '/(\$[a-zA-Z0-9_]+)/'                                                 => "\e[1;31m\$1\e[0m",
+      '/\'([^\']+)\'/'                                                      => "\e[32m'\$1'\e[0m"
     ]));
     return 0;
   }